Monkeypatcher: dont try to find free memory before
While running the mprotect_none test on Ubuntu 24.10 / 24.04 it was
noticed that test assumed that no other program was mmaping/mprotecting
stuff in the area the test was experimenting with. This is a fair
assumption because the program is a single threaded program. Little that
it knew that the last mprotect call for 4*page_size() was mprotecting
a software counters stub area that was created while the program was
running !

By only finding free memory in the high memory area (rather than the low
memory areas where programs usually reside) the chance of this happening
should reduce ??

Relative jumps on x86-64 can be 32bit so only allowing forward jumps to
a (free) stub area should not be a problem. On aarch64 relative jumps
dont go so far so only restricting to free memory after (and not before)
could prove to be problematic