When searching for space to create jump stubs area first search *after* and then *before* memory location (reverse the order)
On x86-64 the max relative jump is +-2GB which should be enough to
find some free space always; in that case it should be found *after*
a particular memory location to nevertheless avoid the potential for
interference with the lower memory regions of the program.

On aarch64 the max relative jump is +-128MB which is much lesser and it
is likely that both branches before and after will be used. Use after
as it has less potential for interference with rest of the program.
Finding memory space before is also fine but is just better to try
the after case first.